add optimized min_distance loop.
Instead of using an arbitrary min_distance in the last stage of Eye.get_ommatidia, we replaced this with an algorithm that seeks to minimize the standard deviation of distances between ommatidial centers. This should, in principle, find the set of points that are most regularly seperated. Requires testing with real data.